+++ src/java/iced-x86/src/main/java/com/github/icedland/iced/x86/fmt/nasm/NasmFormatter.java
@@ -700,7 +700,7 @@ public final class NasmFormatter extends Formatter {
 						numberFormatter, numberOptions, imm8 & 0xFF, symbol, options.getShowSymbolAddress());
 			else {
 				if (numberOptions.signedNumber) {
-					imm64 = (byte)imm8;
+					imm64 = imm8;
 					numberKind = NumberKind.INT8;
 					if (imm8 < 0) {
 						output.write("-", FormatterTextKind.OPERATOR);
@@ -735,7 +735,7 @@ public final class NasmFormatter extends Formatter {
 						numberFormatter, numberOptions, imm16 & 0xFFFF, symbol, options.getShowSymbolAddress());
 			else {
 				if (numberOptions.signedNumber) {
-					imm64 = (short)imm16;
+					imm64 = imm16;
 					numberKind = NumberKind.INT16;
 					if (imm16 < 0) {
 						output.write("-", FormatterTextKind.OPERATOR);
@@ -773,7 +773,7 @@ public final class NasmFormatter extends Formatter {
 				if (numberOptions.signedNumber) {
 					imm64 = imm32;
 					numberKind = NumberKind.INT32;
-					if ((int)imm32 < 0) {
+					if (imm32 < 0) {
 						output.write("-", FormatterTextKind.OPERATOR);
 						imm32 = -imm32;
 					}
